pyproject.toml
# pyproject.toml

[build-system]
requires      = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "weave-py-api"
version = "1.2.0"
description = "Weavechain Python API"
readme = "README.md"
authors = [{ name = "Weavechain", email = "support@weavechain.com" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["weavechain", "weave", "api", "layer-0", "data", "sharing", "monetization", "compute", "node"]
dependencies = [
    "pathlib",
    "requests",
    "base58",
    "pynacl",
    "uuid",
    "websocket-client",
    "cryptography",
    "PyCryptodome",
    "paramiko",
    "java-random",
    'tomli; python_version < "3.11"',
]
requires-python = ">=3.5"

[project.optional-dependencies]
dev = ["black", "bumpver", "isort", "pip-tools", "pytest"]

[project.urls]
Homepage = "https://github.com/weavechain/weave-py-api"

[project.scripts]

[tool.bumpver]
current_version = "1.2.0"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message  = "Bump version {old_version} -> {new_version}"
commit          = true
tag             = true
push            = false

[tool.bumpver.file_patterns]
"pyproject.toml" = ['current_version = "{version}"', 'version = "{version}"']
"src/reader/__init__.py" = ["{version}"]
